Update index.html
it looks like alot of changes but it was made to have consistent spacing and  changes were Only hash files that exist within the manifest. Filenames and hash are now matched. Status will show mismatch properly now if theres an issue. and if theres any mismatch itll show as the screenshot says progressbar will stay open to let the user know how many files are mismatched